chore: move code to shared lib
