+++ apps/slackbot-api/src/app/core/config.ts
@@ -1,0 +1,3 @@
+const slackConfig: any = {}; // TODO: refactor later
+export const SLACK_OAUTH_ACCESS_TOKEN = slackConfig ? slackConfig.access_token : process.env['SLACK_OAUTH_ACCESS_TOKEN'];
+export const SLACK_DEDICATED_CHANNEL = slackConfig ? slackConfig.dedicated_channel : process.env['SLACK_DEDICATED_CHANNEL'];
