+++ libs/api-core/common/src/index.ts
@@ -1,3 +1,4 @@
-export * from './lib/api-core-common.controller';
-export * from './lib/api-core-common.service';
-export * from './lib/api-core-common.module';
+export * from './lib/common.module';
+export * from './lib/models';
+export * from './lib/services';
+export * from './lib/utils';
