--- libs/api-core/common/src/lib/api-core-common.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { ApiCoreCommonService } from './api-core-common.service';
-
-@Controller('api-core-common')
-export class ApiCoreCommonController {
-  constructor(private apiCoreCommonService: ApiCoreCommonService) {}
-}
