--- libs/api-core/common/src/lib/api-core-common.service.ts
@@ -1,4 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class ApiCoreCommonService {}
