+++ libs/api-core/common/src/lib/common.module.ts
@@ -1,0 +1,10 @@
+import { DataModule } from '@foosball/data';
+import { Module } from '@nestjs/common';
+import { DataMartService, MatchService, PlayerService, StatsService } from './services';
+
+@Module({
+  imports: [DataModule],
+  providers: [PlayerService, MatchService, StatsService, DataMartService],
+  exports: [PlayerService, MatchService, StatsService, DataMartService],
+})
+export class ApiCommonModule {}
