+++ libs/api-core/common/src/lib/models/common.slack.ts
