+++ libs/api-core/common/src/lib/models/document.ts
