+++ libs/api-core/common/src/lib/models/index.ts
@@ -1,6 +1,4 @@
 export { IMetrics } from './metrics';
 export { IMatchResult, MatchResult, IMatchResultInput } from './match-result';
-export { IPlayer } from '@foosball/dto';
 export { Callback, IOption, IOptions } from './common.slack';
-
-export type IFinalScore = [number, number];
+export * from './player';
