+++ libs/api-core/common/src/lib/models/time-dimension.ts
@@ -1,4 +1,3 @@
-// import * as moment from 'moment';
 import { DateTime } from 'luxon';
 
 export class TimeDimension {
