+++ libs/api-core/common/src/lib/services/abstract-service.ts
