+++ libs/api-core/common/src/lib/services/player.service.ts
