+++ libs/api-core/common/src/lib/utils/serializers/string-array.serializer.ts
