+++ libs/api-core/common/tsconfig.lib.json
@@ -10,6 +10,6 @@
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["**/*.ts"],
+  "include": ["**/*.ts", "../../../apps/slackbot-api/src/app/slack-helper.ts"],
   "exclude": ["jest.config.ts", "**/*.spec.ts", "**/*.test.ts"]
 }
