+++ libs/shared/dto/src/index.ts
@@ -1,3 +1,3 @@
 // export * from './lib/enums';
 export * from './lib/interfaces';
-// export * from './lib/types';
+export * from './lib/types';
