+++ libs/shared/dto/src/lib/interfaces/cube-item.interface.ts
