+++ libs/shared/dto/src/lib/interfaces/index.ts
@@ -1,1 +1,3 @@
+export * from './cube-item.interface';
+export * from './match-cube-item.interface';
 export * from './player.interface';
