+++ libs/shared/dto/src/lib/interfaces/match-cube-item.interface.ts
