+++ libs/shared/dto/src/lib/types/final-score.type.ts
@@ -1,0 +1,1 @@
+export type IFinalScore = [number, number];
