+++ libs/shared/dto/src/lib/types/index.ts
@@ -1,0 +1,1 @@
+export * from './final-score.type';
